rename packages to match intent
