+++ src/org/opensourcebim/ifccollection/IfcToMpgCollectionService.java
@@ -1,4 +1,4 @@
-package org.opensourcebim.mpgcalculations;
+package org.opensourcebim.ifccollection;
 
 import org.bimserver.bimbots.BimBotContext;
 import org.bimserver.bimbots.BimBotsException;
@@ -7,8 +7,12 @@ import org.bimserver.bimbots.BimBotsOutput;
 import org.bimserver.emf.IfcModelInterface;
 import org.bimserver.interfaces.objects.SObjectType;
 import org.bimserver.plugins.services.BimBotAbstractService;
+import org.opensourcebim.nmd.BimMaterialDatabaseSession;
+import org.opensourcebim.nmd.NmdDataBaseSession;
+import org.opensourcebim.nmd.NmdDataResolver;
+import org.opensourcebim.nmd.NmdDataResolverImpl;
 
-public class MpgCalculator extends BimBotAbstractService {
+public class IfcToMpgCollectionService extends BimBotAbstractService {
 
 	@Override
 	public BimBotsOutput runBimBot(BimBotsInput input, BimBotContext bimBotContext, SObjectType settings)
@@ -20,9 +24,23 @@ public class MpgCalculator extends BimBotAbstractService {
 		MpgIfcObjectCollector matParser = new MpgIfcObjectCollector();
 		MpgObjectStore ifcResults = matParser.collectIfcModelObjects(ifcModel);
 		
+		// notify user of any warnings:
+		if(ifcResults.CheckForWarningsAndErrors()) {
+			// return list of warnings
+		} else {
+			// return model has been read correctly.
+		}
+		
 		// find matching material properties from Material DB
+		NmdDataResolver nmdDataProvider = new NmdDataResolverImpl();
+		nmdDataProvider.addService(new NmdDataBaseSession());
+		nmdDataProvider.addService(new BimMaterialDatabaseSession());
+		
+		
+		MpgObjectStore nmdResults = nmdDataProvider.NmdToMpg(ifcResults);
 		
 		// notify user of unknown materials
+		
 		
 		// retrieve user input
